NBA Shotmap Bot/requirements.txt
discord.py
nba_api
matplotlib
pandas
numpy
requests
